+++ libTidyEngine/mesh.cpp
@@ -24,7 +24,7 @@ Mesh::Mesh()
 	;
 }
 
-Mesh::Mesh(std::vector<Vertex> &vertices, std::vector<uint64_t> &indices, std::vector<GLuint> &textures)
+Mesh::Mesh(std::vector<Vertex> &vertices, std::vector<uint32_t> &indices, std::vector<GLuint> &textures)
 {
 	m_Vertices = vertices;
 	m_Indices = indices;
